refactor: break out into multiple components
