--- public/request/user/list-currently-active.php
@@ -1,5 +1,0 @@
-<?php
-
-return response()->json(
-    getCachedLatestRichPresenceUpdates()
-);
