--- resources/css/active-players.css
@@ -1,18 +1,0 @@
-#activeplayersbox {
-  min-height: 25.0em;
-  height: 25.0em;
-  max-height: 25.0em;
-  overflow-y: auto;
-}
-
-#active-players-filter-options {
-  margin-top: 3px;
-}
-
-#active-players-menu-button::before {
-  content: '☰';
-}
-
-#active-players-menu-button.menuOpen::before {
-  content: '✕';
-}
