+++ resources/css/app.css
@@ -38,7 +38,6 @@
 @import "tooltip.css";
 
 /* TODO split into components or replace with utility classes */
-@import "active-players.css";
 @import "award-count.css";
 @import "comment.css";
 @import "devbox.css";
